fix(ci): second review pass — before-checkout rollback gate, resolver reason, file limits
Independent two-axis review (code-review skill) findings, fixed:

Spec:
- rollback.yml resolves manifest eligibility BEFORE checkout via the GitHub
  API with inline pinned-identity verification (blob id + SHA-256), so
  rollback-ineligible components stop before checkout, dependency install,
  or Wrangler; checkout/install/List-versions only run when a future
  owner-approved manifest marks a component eligible.
- resolver now emits a typed non-sensitive eligibility.reason (candidate
  match / stale / deploy-ineligible component / no candidate).
- production Atlas apply fails closed on empty NEON_DATABASE_URL instead of
  silently skipping (staging keeps the skip).

Standards (1-10-50):
- validation constants + component validators extracted to
  release-manifest-validation.rb (resolver 306 -> 131 lines; no 4-level
  nesting; validators <= 10 lines).
- test files trimmed to the <=200-line limit
  (release-manifest-resolver.test.rb 214 -> 179; test_safe1 -> 200).

Written by opencode (opencode-go/deepseek-v4-flash, variant max)
