+++ .github/scripts/release-manifest-resolver.test.rb
@@ -51,160 +51,125 @@ def mutate_manifest(copy_path, &block)
   File.write(copy_path, JSON.generate(manifest))
 end
 
+
+def with_copy(name)
+  Dir.mktmpdir(name) do |dir|
+    copy = File.join(dir, "manifest.json")
+    FileUtils.cp(MANIFEST, copy)
+    yield copy
+  end
+end
+
 # ── Green: the real pinned manifest, every component ─────────────────────────
 %w[catalog web users maintenance root].each do |component|
   green_case("pinned manifest resolves #{component}", MANIFEST, component)
 end
 
-# ── Green: eligibility verdict — matching revision → deploy true ────────────
+# ── Green: eligibility verdicts ──────────────────────────────────────────────
 green = green_case("matching revision is deploy-eligible", MANIFEST, "root")
 abort "FAIL: matching revision must be deploy eligible" unless green.dig("eligibility", "deploy") == true
 abort "FAIL: rollback must be ineligible per manifest policy" unless green.dig("eligibility", "rollback") == false
 
-# ── Green: non-matching revision → deploy ineligible (typed, not an error) ──
 stale = green_case("stale revision is deploy-ineligible", MANIFEST, "root", "0" * 40)
 abort "FAIL: stale revision must be deploy ineligible" unless stale.dig("eligibility", "deploy") == false
 
-# ── Green: no revision argument → eligibility.deploy omitted ────────────────
 no_rev = green_case("revision omitted omits deploy verdict", MANIFEST, "catalog", nil)
 abort "FAIL: omitted revision must omit deploy verdict" unless no_rev["eligibility"].key?("deploy") && no_rev.dig("eligibility", "deploy").nil?
 
 # ── Red: tampered manifest — altering any field fails the pinned identity ───
-Dir.mktmpdir("b1-tamper") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-tamper") do |copy|
   mutate_manifest(copy) { |m| m["atlas"]["target"] = "20260909000000" }
-  red_case("altered atlas target fails pinned identity", copy, "root",
-           ["manifest SHA-256", "atlas.target must be pinned 20260809000031"])
+  red_case("altered atlas target fails pinned identity", copy, "root", ["manifest SHA-256", "atlas.target must be pinned 20260809000031"])
 end
 
 # ── Red: unknown top-level field ────────────────────────────────────────────
-Dir.mktmpdir("b1-unknown-top") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-unknown-top") do |copy|
   mutate_manifest(copy) { |m| m["surprise"] = true }
-  red_case("unknown top-level field", copy, "root",
-           ["manifest SHA-256", "unknown top-level field(s): surprise"])
+  red_case("unknown top-level field", copy, "root", ["manifest SHA-256", "unknown top-level field(s): surprise"])
 end
 
 # ── Red: unknown component key ──────────────────────────────────────────────
-Dir.mktmpdir("b1-unknown-comp") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-unknown-comp") do |copy|
   mutate_manifest(copy) { |m| m["components"]["sidecar"] = m["components"]["root"] }
-  red_case("unknown component key", copy, "root",
-           ["manifest SHA-256", "unknown component(s): sidecar"])
+  red_case("unknown component key", copy, "root", ["manifest SHA-256", "unknown component(s): sidecar"])
 end
 
 # ── Red: requested component not in manifest ────────────────────────────────
-red_case("unknown requested component", MANIFEST, "sidecar",
-         ["requested component \"sidecar\" is not in the manifest"])
+red_case("unknown requested component", MANIFEST, "sidecar", ["requested component \"sidecar\" is not in the manifest"])
 
 # ── Red: schema_version must be exactly 1 ───────────────────────────────────
-Dir.mktmpdir("b1-schema") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-schema") do |copy|
   mutate_manifest(copy) { |m| m["schema_version"] = 2 }
-  red_case("schema_version 2 rejected", copy, "root",
-           ["manifest SHA-256", "schema_version must be 1, got 2"])
+  red_case("schema_version 2 rejected", copy, "root", ["manifest SHA-256", "schema_version must be 1, got 2"])
 end
 
 # ── Red: source_revision must be a full 40-hex SHA ──────────────────────────
-Dir.mktmpdir("b1-rev") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-rev") do |copy|
   mutate_manifest(copy) { |m| m["source_revision"] = "deadbeef" }
-  red_case("short source_revision rejected", copy, "root",
-           ["manifest SHA-256", "source_revision must be a full 40-hex SHA"])
+  red_case("short source_revision rejected", copy, "root", ["manifest SHA-256", "source_revision must be a full 40-hex SHA"])
 end
 
 # ── Red: atlas digest must match the pinned digest exactly ──────────────────
-Dir.mktmpdir("b1-digest") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-digest") do |copy|
   mutate_manifest(copy) { |m| m["atlas"]["digest_sha256"] = "e" * 64 }
-  red_case("altered atlas digest rejected", copy, "root",
-           ["manifest SHA-256", "atlas.digest_sha256 must be pinned"])
+  red_case("altered atlas digest rejected", copy, "root", ["manifest SHA-256", "atlas.digest_sha256 must be pinned"])
 end
 
 # ── Red: atlas directory must be migrations/neon ────────────────────────────
-Dir.mktmpdir("b1-atlas-dir") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-atlas-dir") do |copy|
   mutate_manifest(copy) { |m| m["atlas"]["directory"] = "db/migrations" }
-  red_case("altered atlas directory rejected", copy, "root",
-           ["manifest SHA-256", "atlas.directory must be migrations/neon"])
+  red_case("altered atlas directory rejected", copy, "root", ["manifest SHA-256", "atlas.directory must be migrations/neon"])
 end
 
 # ── Red: non-boolean eligibility ────────────────────────────────────────────
-Dir.mktmpdir("b1-elig") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-elig") do |copy|
   mutate_manifest(copy) { |m| m["components"]["root"]["rollback_eligible"] = "maybe" }
-  red_case("non-boolean rollback_eligible rejected", copy, "root",
-           ["manifest SHA-256", "component root: rollback_eligible must be a boolean"])
+  red_case("non-boolean rollback_eligible rejected", copy, "root", ["manifest SHA-256", "component root: rollback_eligible must be a boolean"])
 end
 
 # ── Red: invalid secret name ────────────────────────────────────────────────
-Dir.mktmpdir("b1-secret") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-secret") do |copy|
   mutate_manifest(copy) { |m| m["components"]["catalog"]["worker_secrets"] = ["bad-name!"] }
-  red_case("invalid secret name rejected", copy, "root",
-           ["manifest SHA-256", "component catalog: secret name \"bad-name!\" must match"])
+  red_case("invalid secret name rejected", copy, "root", ["manifest SHA-256", "component catalog: secret name \"bad-name!\" must match"])
 end
 
 # ── Red: depends_on referencing an unknown component ────────────────────────
-Dir.mktmpdir("b1-dep") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-dep") do |copy|
   mutate_manifest(copy) { |m| m["components"]["root"]["depends_on"] = ["phantom"] }
-  red_case("depends_on unknown component rejected", copy, "root",
-           ["manifest SHA-256", "component root: depends_on \"phantom\" is not a known component"])
+  red_case("depends_on unknown component rejected", copy, "root", ["manifest SHA-256", "component root: depends_on \"phantom\" is not a known component"])
 end
 
 # ── Red: component drift from the expected surface (worker name) ────────────
-Dir.mktmpdir("b1-drift") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-drift") do |copy|
   mutate_manifest(copy) { |m| m["components"]["web"]["worker_name"] = "animichi-web-renamed" }
-  red_case("worker-name drift rejected", copy, "root",
-           ["manifest SHA-256", "component web: worker_name must be \"animichi-web\""])
+  red_case("worker-name drift rejected", copy, "root", ["manifest SHA-256", "component web: worker_name must be \"animichi-web\""])
 end
 
 # ── Red: missing component ──────────────────────────────────────────────────
-Dir.mktmpdir("b1-missing") do |dir|
-  copy = File.join(dir, "manifest.json")
-  FileUtils.cp(MANIFEST, copy)
+with_copy("b1-missing") do |copy|
   mutate_manifest(copy) { |m| m["components"].delete("maintenance") }
-  red_case("missing component rejected", copy, "root",
-           ["manifest SHA-256", "missing component(s): maintenance"])
+  red_case("missing component rejected", copy, "root", ["manifest SHA-256", "missing component(s): maintenance"])
 end
 
 # ── Red: manifest must be valid JSON ────────────────────────────────────────
 Dir.mktmpdir("b1-json") do |dir|
   copy = File.join(dir, "manifest.json")
   File.write(copy, "{ not json")
-  red_case("invalid JSON rejected", copy, "root",
-           ["manifest is not valid JSON"])
+  red_case("invalid JSON rejected", copy, "root", ["manifest is not valid JSON"])
 end
 
 # ── Red: component marked deploy_eligible=false stays ineligible even with a
-#    matching revision (the flag is enforced, not just validated) ────────────
-# NOTE: the pinned-manifest path cannot hold a deploy_eligible=false component
-# (the pinned blob is immutable), so this behavior is verified structurally
-# here: the resolver's eligibility.deploy must AND the flag with the revision
-# match. Asserted via source inspection rather than a fixture (a tampered copy
-# fails on the pinned identity before reaching the verdict).
+#    matching revision. The pinned blob cannot hold such a component, so this
+#    is verified structurally: eligibility.deploy must AND the flag with the
+#    revision match (a tampered copy fails on the pinned identity first).
 resolver_source = File.read(SCRIPT)
-unless resolver_source.include?("SOURCE_REVISION == manifest.fetch(\"source_revision\") && component.fetch(\"deploy_eligible\")")
+unless resolver_source.include?("revision_matches && component.fetch(\"deploy_eligible\")")
   abort "FAIL: resolver must enforce component.deploy_eligible in the deploy verdict"
 end
 puts "PASS: resolver enforces component.deploy_eligible in the deploy verdict"
 
-# ── Green: manifest_blob_id in the verdict is the COMPUTED blob id (observed
-#    value), so consumers never mistake an unavailable computation for a
-#    verified one ─────────────────────────────────────────────────────────────
+# ── Green: manifest_blob_id is the COMPUTED blob id (never mistaken for a
+#    verified constant when the computation was unavailable) ─────────────────
 computed = `git -C "#{ROOT}" hash-object .github/release-manifests/production-pre-campaign.json`.strip
 parsed = green_case("computed blob id reported", MANIFEST, "root")
 abort "FAIL: manifest_blob_id must be the computed #{computed}, got #{parsed['manifest_blob_id'].inspect}" \
