+++ .github/workflows/rollback.yml
@@ -31,12 +31,54 @@ jobs:
     timeout-minutes: 10
     steps:
       # SAFE-1 (#937): the caller-supplied `version_id` input is removed. The
-      # component/config mapping and rollback eligibility now resolve from the
-      # pinned production manifest (content-addressed GitHub blob — never the
-      # working tree). The manifest represents rollback as ineligible today,
-      # so every rollback stops here, before any Wrangler command; only a
-      # future owner-approved manifest revision can mark a component
-      # rollback-eligible with an exact version pair.
+      # component/config mapping and rollback eligibility resolve from the
+      # pinned production manifest — read content-addressed from the GitHub
+      # API here, WITHOUT checking out any source, so rollback-ineligible
+      # components stop before checkout, dependency install, or Wrangler. The
+      # manifest represents rollback as ineligible today, so every rollback
+      # stops at this first step; only a future owner-approved manifest
+      # revision can mark a component rollback-eligible with an exact version
+      # pair.
+      - name: Resolve manifest rollback eligibility
+        id: resolve
+        shell: bash
+        env:
+          COMPONENT: ${{ inputs.component }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          PINNED_MANIFEST_BLOB_ID: 2aef389878e321ddc3d4d86922fde30ad2ab6491
+          PINNED_MANIFEST_SHA256: 1ffcfe1bf4815e0e0e890193cbfdeb28d9748f94bb98db88106000af708f536c
+        run: |
+          set -euo pipefail
+          tmp="$(mktemp -d)"
+          trap 'rm -rf "$tmp"' EXIT
+          gh api "repos/${GITHUB_REPOSITORY}/git/blobs/${PINNED_MANIFEST_BLOB_ID}" \
+            --jq .content | base64 -d > "$tmp/manifest.json"
+          # Inline pinned-identity check (mirrors release-manifest-resolver.rb;
+          # the contract test asserts these pins stay in sync). Never trust a
+          # checkout copy of the manifest here.
+          actual_sha="$(shasum -a 256 "$tmp/manifest.json" | awk '{print $1}')"
+          if [ "$actual_sha" != "$PINNED_MANIFEST_SHA256" ]; then
+            echo "::error::pinned manifest SHA-256 mismatch ($actual_sha) — refusing any rollback"
+            exit 1
+          fi
+          verdict="$(COMPONENT="$COMPONENT" python3 - "$tmp/manifest.json" <<'PY'
+          import json, os, sys
+          manifest = json.load(open(sys.argv[1]))
+          comp = manifest["components"][os.environ["COMPONENT"]]
+          print(f"{comp['rollback_eligible']} {comp['working_directory']}")
+          PY
+          )"
+          eligible="${verdict%% *}"
+          directory="${verdict#* }"
+          if [ "$eligible" != "True" ] && [ "$eligible" != "true" ]; then
+            echo "::error::rollback-ineligible: the pinned SAFE-1 manifest marks component '$COMPONENT' rollback-ineligible (no approved component/version pair exists). No checkout or Wrangler rollback was attempted."
+            exit 1
+          fi
+          echo "working_directory=$directory" >> "$GITHUB_OUTPUT"
+          echo "config_flag=$([ "$COMPONENT" = "root" ] && echo '-c workers/edge/wrangler.toml' || echo '')" >> "$GITHUB_OUTPUT"
+      # Reachable only after a future owner-approved manifest marks a
+      # component rollback-eligible with an exact version pair; the current
+      # manifest cannot reach these steps.
       - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
         with:
           persist-credentials: false
@@ -50,25 +92,6 @@ jobs:
           cache-dependency-path: pnpm-lock.yaml
       - name: Install dependencies
         run: pnpm install --frozen-lockfile --ignore-scripts
-      - name: Resolve manifest rollback eligibility
-        id: resolve
-        shell: bash
-        env:
-          COMPONENT: ${{ inputs.component }}
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-          set -euo pipefail
-          verdict="$(bash .github/scripts/release-eligibility.sh "$COMPONENT" "${{ github.sha }}")"
-          eligible="$(echo "$verdict" | jq -r .component.rollback_eligible)"
-          if [ "$eligible" != "true" ]; then
-            echo "::error::rollback-ineligible: the pinned SAFE-1 manifest marks component '$COMPONENT' rollback-ineligible (no approved component/version pair exists). No Wrangler rollback was attempted."
-            exit 1
-          fi
-          echo "working_directory=$(echo "$verdict" | jq -r .component.working_directory)" >> "$GITHUB_OUTPUT"
-          echo "config_flag=$([ "$COMPONENT" = "root" ] && echo '-c workers/edge/wrangler.toml' || echo '')" >> "$GITHUB_OUTPUT"
-      # Reachable only after a future owner-approved manifest marks a
-      # component rollback-eligible with an exact version pair; the current
-      # manifest cannot reach these steps.
       - name: List versions
         working-directory: ${{ steps.resolve.outputs.working_directory }}
         env:
